+++ packages/form-schema/src/core/index.ts
@@ -1,4 +1,5 @@
 export { serialize, defaultSerializer } from "./serialize";
+export { serializeSingle } from "./serializeSingle";
 export { parse, defaultParser } from "./parse";
 export { setFieldErrors, defaultFieldErrorsSetter } from "./setFieldErrors";
 export { validateBeforeSubmit } from "./validateBeforeSubmit";