+++ packages/form-schema/src/fields/checkbox/checkbox.ts
@@ -6,13 +6,7 @@ export function prepareValue(rawValue: unknown) {
 }
 
 export const checkbox: FieldType<CheckboxSchema> = {
-	serializer: ({ name, values }) => {
-		const value = values[name];
-
-		return {
-			[name]: prepareValue(value),
-		};
-	},
+	serializerSingle: ({ value }) => prepareValue(value),
 
 	parser: ({ name, values }) => {
 		const value = values[name];