+++ packages/form-schema/src/fields/list/schema.ts
@@ -1,0 +1,5 @@
+export type ListSchema<FieldSchema> = {
+  maxLength?: number;
+	minLength?: number;
+  itemSchema: FieldSchema;
+};